pyproject.toml
[tool.poetry]
name = "stable-bot"
version = "0.1.0"
description = ""
authors = ["Royal Simpson Pinto <royalpinto007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
replicate = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
